+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -20,8 +20,8 @@ import java.util.Map;
 @RequestMapping("/users")
 public class UserController {
 
-    private final Map<Integer, User> users = new HashMap<>();
-    private int id = 1;
+    private final Map<Long, User> users = new HashMap<>();
+    private Long id = 1L;
 
     /**
      * Создание нового пользователя.
@@ -100,7 +100,7 @@ public class UserController {
      *
      * @return Id пользователя.
      */
-    private int generateId() {
+    private Long generateId() {
         return id++;
     }
 }
